GeoffLethbridge/Marlin v1.0.2 – January 2015 (CNC Mini Mill at CCHS) forked from MarlinFirmware/Marlin branch Marlin_v1.0.2
The move to Marlin_v1.0.2 is required for the use of the new LCD.

The old Marlin v1 for the Mini Mill has been scraped as there are too
many undocumented changes.

There were are a large number of changes made off github to he old
Marlin v1 for the Mini Mill. rdpowers/Marlin forked from
MarlinFirmware/Marlin branch Marlin_v1. this was merged with a group of
changes by RobG that that was made to a version of Marlin_v1 that was on
the CCHS networked drive. This merge revealed 22 files that had changes,
and all being undocumented changes.